paasio: simplify tests (#1982)
[no important files changed]

part of https://github.com/exercism/rust/issues/1824

This is by no means perfect, but I attempted to make the test functions
a little more concise and readable. This is desirable when the test code
is shown to users online, while it wasn't a requirement when these tests
were generated by macros.