palindrome-products: require factors in solution (#2006)
There were two extra tests on the now-defunct `Palindrome::new`
function. I felt those only made sense if this exercise still contained
the newtype pattern.

The decision to return a HashSet was done to communicate to the student
that order does not matter. The order of the items inside the tuple pair
is the same as the description of the exercise.

An owned HashSet was chosen to allow for solutions that don't store a
HashSet in Palindrome (or elsewhere) and would have trouble producing a
reference to one. See the feedback:
https://github.com/exercism/rust/pull/2006?#discussion_r1935384069

Corresponding forum post, which goes into more detail about why this
change was necessary:
https://forum.exercism.org/t/palindrome-products-description-mentions-factors/

---------

Co-authored-by: Remo Senekowitsch <remo@buenzli.dev>