Improve test no_underscore_prefixed_idents (#2078)
* accumulate doesn't need to be in the exceptions list because it
doesn't compile independent of the argument name and it is allowed to do
so.

* xorcism doesn't need to be in the exceptions list because the compiler
doesn't actually complain about the unused field if the name doesn't
have a prefix. This only applies to fields of type `PhantomData`. I
believe this must be a recent addition. It's nice, because the compiler
can know that `PhantomData` is obviously never read.

* Fix some typos in the documentation of the test.