Do not hide test output in log file (#2100)
Putting the clippy output in a log file is somewhat convenient for local
use. But if there's a failure in CI that's difficult to reproduce
locally, it only makes things harder to debug.

The actual problem must've been that clippy used to be installed in the
GitHub Actions environment by default (but not anymore), so we didn't
notice that it's not explicitly requested in our workflow.